requirements.txt
Flask==3.1.0
python-jose==3.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
Flask-SQLAlchemy==3.1.1
qrcode==8.0
pillow==11.1.0
Flask-Login==0.6.3
requests-oauthlib==2.0.0